fix: tests

+++ src/main/java/ru/yandex/practicum/filmorate/storage/repository/db/FilmStorageImpl.java
@@ -214,10 +214,12 @@ public class FilmStorageImpl implements FilmStorage {
         }
 
         String insertLikesSql = "insert into likes(film_id, user_id) values (?, ?)";
-        if (!film.getLikes().isEmpty()) {
-            for (Long userWhoLikeId : film.getLikes()) {
-                if (userWhoLikeId > 0) {
-                    jdbcTemplate.update(insertLikesSql, film.getId(), userWhoLikeId);
+        if (film.getLikes() != null) {
+            if (!film.getLikes().isEmpty()) {
+                for (Long userWhoLikeId : film.getLikes()) {
+                    if (userWhoLikeId > 0) {
+                        jdbcTemplate.update(insertLikesSql, film.getId(), userWhoLikeId);
+                    }
                 }
             }
         } else {
@@ -232,26 +234,28 @@ public class FilmStorageImpl implements FilmStorage {
         }
 
         String insertGenresSql = "insert into film_genre(film_id, genre_id) values (?, ?)";
-        for (Film.Genre genre : film.getGenres()) {
-            jdbcTemplate.update(insertGenresSql, film.getId(), genre.getId());
-        }
+        if (film.getGenres() != null) {
+            for (Film.Genre genre : film.getGenres()) {
+                jdbcTemplate.update(insertGenresSql, film.getId(), genre.getId());
+            }
 
-        if (selectFilmGenre().get(film.getId()) != null) {
-            film.setGenres(selectFilmGenre().get(film.getId()));
+            if (selectFilmGenre().get(film.getId()) != null) {
+                film.setGenres(selectFilmGenre().get(film.getId()));
+            }
         }
-
     }
 
     private void insertFilmMpa(Film film, boolean update) {
-        if (update) {
-            String deleteMpaSql = "delete from mpa where film_id=?";
-            jdbcTemplate.update(deleteMpaSql, film.getId());
+        if (film.getMpa() != null) {
+            if (update) {
+                String deleteMpaSql = "delete from mpa where film_id=?";
+                jdbcTemplate.update(deleteMpaSql, film.getId());
+            }
+
+            String insertMpaSql = "insert into mpa(film_id, mpa_id) values (?, ?)";
+
+            jdbcTemplate.update(insertMpaSql, film.getId(), film.getMpa().getId());
         }
-
-        String insertMpaSql = "insert into mpa(film_id, mpa_id) values (?, ?)";
-
-        jdbcTemplate.update(insertMpaSql, film.getId(), film.getMpa().getId());
-
     }
 
     private Map<Long, Set<Long>> selectFilmLikes() {
